+++ frontend/src/mocks/fixtures/api/projects/:team_id/insights/trendsAreaBreakdown.json
