+++ frontend/src/scenes/insights/EmptyStates/EmptyStates.stories.tsx
@@ -3,7 +3,7 @@ import { Meta } from '@storybook/react'
 import funnelOneStep from './funnelOneStep.json'
 import { useStorybookMocks } from '~/mocks/browser'
 import { router } from 'kea-router'
-import insight from '../__mocks__/trendsLine.json'
+import insight from '../../../mocks/fixtures/api/projects/:team_id/insights/trendsLine.json'
 import { insightLogic } from 'scenes/insights/insightLogic'
 import { InsightShortId } from '~/types'
 import { createInsightStory } from 'scenes/insights/__mocks__/createInsightScene'
